+++ CHANGELOG.md
@@ -8,8 +8,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 ### Added
-- Layer integration tests for more apps. Closes
-  [[#472](https://github.com/metalbear-co/mirrord/issues/472)].
+- Layer integration tests with more apps (testing with Go only on MacOS because of
+  known crash on Linux - [[#380](https://github.com/metalbear-co/mirrord/issues/380)]).
+  Closes [[#472](https://github.com/metalbear-co/mirrord/issues/472)].
 
 ## Changed
 - Don't report InProgress io error as error (log as info)